use root document to include copyrights
