+++ src/main/java/com/graphhopper/converter/api/GHResponse.java
@@ -1,94 +1,28 @@
 package com.graphhopper.converter.api;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * @author Robin Boldt
+ * @author Peter Karich
  */
 public class GHResponse {
 
-    private Long osmId;
-    private String osmType;
-
-    private Point point;
-
-    private String name;
-    private String country;
-    private String city;
     private List<String> copyrights = new ArrayList<String>(5);
+    private List<GHEntry> hits;
 
-    public GHResponse(Long osmId, String type, double lat, double lng, String name, String country, String city) {
-        this.osmId = osmId;
-        this.osmType = type;
-        this.point = new Point(lat, lng);
-        this.name = name;
-        this.country = country;
-        this.city = city;
+    public GHResponse() {
+        this(5);
     }
 
-    @JsonProperty
-    public String getName() {
-        return name;
+    public GHResponse(int no) {
+        hits = new ArrayList<>(no);
     }
 
-    @JsonProperty
-    public void setName(String name) {
-        this.name = name;
+    public void setCopyrights(List<String> copyrights) {
+        this.copyrights = copyrights;
     }
 
-    @JsonProperty
-    public String getCountry() {
-        return country;
-    }
-
-    @JsonProperty
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    @JsonProperty
-    public String getCity() {
-        return city;
-    }
-
-    @JsonProperty
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    @JsonProperty
-    public Point getPoint() {
-        return point;
-    }
-
-    @JsonProperty
-    public void setPoint(Point point) {
-        this.point = point;
-    }
-
-    @JsonProperty("osm_id")
-    public Long getOsmId() {
-        return osmId;
-    }
-
-    @JsonProperty("osm_id")
-    public void setOsmId(Long osmId) {
-        this.osmId = osmId;
-    }
-
-    @JsonProperty("osm_type")
-    public String getOsmType() {
-        return osmType;
-    }
-
-    @JsonProperty("osm_type")
-    public void setOsmType(String type) {
-        this.osmType = type;
-    }
-
-    @JsonProperty("copyrights")
     public List<String> getCopyrights() {
         return copyrights;
     }
@@ -98,40 +32,15 @@ public class GHResponse {
         return this;
     }
 
-    @JsonProperty("copyrights")
-    public void setCopyrights(List<String> copyrights) {
-        this.copyrights = copyrights;
+    public void setHits(List<GHEntry> hits) {
+        this.hits = hits;
     }
 
-    public class Point {
-
-        private double lat;
-        private double lng;
-
-        public Point(double lat, double lng) {
-            this.lat = lat;
-            this.lng = lng;
-        }
-
-        @JsonProperty
-        public double getLat() {
-            return lat;
-        }
-
-        @JsonProperty
-        public void setLat(double lat) {
-            this.lat = lat;
-        }
-
-        @JsonProperty
-        public double getLng() {
-            return lng;
-        }
-
-        @JsonProperty
-        public void setLng(double lng) {
-            this.lng = lng;
-        }
+    public void add(GHEntry entry) {
+        hits.add(entry);
     }
 
+    public List<GHEntry> getHits() {
+        return hits;
+    }
 }
